requirements.txt
psycopg2-binary>=2.9
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
scipy>=1.11
pyyaml>=6.0
sqlalchemy>=2.0
